feat: use dot product and limit allowed files to 16b wave files

+++ dtmf/src/fft.h
@@ -11,7 +11,7 @@ typedef float complex cplx_t;
 
 int fft(cplx_t *buf, size_t n);
 
-void float_to_cplx_t(const int32_t *in, cplx_t *out, size_t n);
+void float_to_cplx_t(const int16_t *in, cplx_t *out, size_t n);
 void extract_frequencies(const cplx_t *buf, size_t n, double sample_rate,
 			 uint32_t *f1, uint32_t *f2);
